requirements.txt
click==7.1.2
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.19.1
opencv-python==4.3.0.36
pandas==1.0.5
pickle-mixin==1.0.2
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
